requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
psycopg2==2.9.3
PyJWT==1.7.1
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
zipp==3.7.0
